fix: extract preflight into _preflight.py and fix test failures
Move _check_dispatch_feasibility, _get_fix_required_hook_matchers, and
filter_steps_by_post_prune into server/tools/_preflight.py to satisfy
line-count and business-logic-in-handler arch constraints. Fix test
issues: BackendCapabilities constructor, missing post_prune_step_names
in mocks, MagicMock config_providers crash, missing pytest import,
cascade map entry, and HOOK_REGISTRY monkeypatch targets.
