+++ tests/server/test_tools_fleet_dispatch_preflight.py
@@ -9,6 +9,8 @@ from __future__ import annotations
 
 import inspect
 
+import pytest
+
 pytestmark = [pytest.mark.layer("server"), pytest.mark.small]
