+++ tests/server/test_tools_kitchen_preflight.py
@@ -18,7 +18,11 @@ from unittest.mock import MagicMock, patch
 
 import pytest
 
+from autoskillit.config._config_dataclasses import ProvidersConfig
+
 pytestmark = [pytest.mark.layer("server"), pytest.mark.small]
+
+_DEFAULT_PROVIDERS = ProvidersConfig()
 
 
 def _make_recipe_step(name: str, tool: str | None = "run_skill", **kwargs: Any) -> MagicMock:
@@ -60,7 +64,6 @@ def _make_codex_backend() -> MagicMock:
     from autoskillit.core import BackendCapabilities
 
     caps = BackendCapabilities(
-        name="codex",
         applicable_guards=frozenset(),
         anthropic_provider_capable=False,
     )
@@ -71,12 +74,11 @@ def _make_codex_backend() -> MagicMock:
 
 
 def _make_claude_backend() -> MagicMock:
-    """Create a mock claude-code backend with applicable_guards."""
+    """Create a mock claude-code backend with applicable_guards covering the synthetic hook."""
     from autoskillit.core import AGENT_BACKEND_CLAUDE_CODE, BackendCapabilities
 
     caps = BackendCapabilities(
-        name=AGENT_BACKEND_CLAUDE_CODE,
-        applicable_guards=frozenset({"skill_load_guard"}),
+        applicable_guards=frozenset({"skill_load_guard", "synthetic_test_hook"}),
         anthropic_provider_capable=True,
     )
     backend = MagicMock()
@@ -95,7 +97,7 @@ class TestCheckDispatchFeasibilityUnit:
 
     def test_no_run_skill_steps_returns_none(self) -> None:
         """A recipe with no run_skill steps passes the preflight (1c)."""
-        from autoskillit.server.tools.tools_execution import _check_dispatch_feasibility
+        from autoskillit.server.tools._preflight import _check_dispatch_feasibility
 
         backend = _make_codex_backend()
         active_steps: dict[str, Any] = {
@@ -103,48 +105,48 @@ class TestCheckDispatchFeasibilityUnit:
             "step2": _make_recipe_step("step2", tool="run_python"),
         }
         synthetic = _make_fix_required_hook()
-        with patch("autoskillit.hook_registry.HOOK_REGISTRY", [synthetic]):
+        with patch("autoskillit.server.tools._preflight.HOOK_REGISTRY", [synthetic]):
             result = _check_dispatch_feasibility(
                 post_prune_step_names=["step1", "step2"],
                 active_recipe_steps=active_steps,
                 backend=backend,
-                config_providers=MagicMock(),
+                config_providers=_DEFAULT_PROVIDERS,
             )
         assert result is None, f"Expected None, got: {result}"
 
     def test_compatible_backend_passes(self) -> None:
         """Backend with applicable_guards covering the fix-required hook passes (1e)."""
-        from autoskillit.server.tools.tools_execution import _check_dispatch_feasibility
+        from autoskillit.server.tools._preflight import _check_dispatch_feasibility
 
         backend = _make_claude_backend()
         active_steps: dict[str, Any] = {
             "step1": _make_recipe_step("step1", tool="run_skill"),
         }
         synthetic = _make_fix_required_hook()
-        with patch("autoskillit.hook_registry.HOOK_REGISTRY", [synthetic]):
+        with patch("autoskillit.server.tools._preflight.HOOK_REGISTRY", [synthetic]):
             result = _check_dispatch_feasibility(
                 post_prune_step_names=["step1"],
                 active_recipe_steps=active_steps,
                 backend=backend,
-                config_providers=MagicMock(),
+                config_providers=_DEFAULT_PROVIDERS,
             )
         assert result is None, f"Expected None for compatible backend, got: {result}"
 
     def test_incompatible_backend_fails(self) -> None:
         """Backend with empty applicable_guards returns error (1b)."""
-        from autoskillit.server.tools.tools_execution import _check_dispatch_feasibility
+        from autoskillit.server.tools._preflight import _check_dispatch_feasibility
 
         backend = _make_codex_backend()
         active_steps: dict[str, Any] = {
             "step1": _make_recipe_step("step1", tool="run_skill"),
         }
         synthetic = _make_fix_required_hook()
-        with patch("autoskillit.hook_registry.HOOK_REGISTRY", [synthetic]):
+        with patch("autoskillit.server.tools._preflight.HOOK_REGISTRY", [synthetic]):
             result = _check_dispatch_feasibility(
                 post_prune_step_names=["step1"],
                 active_recipe_steps=active_steps,
                 backend=backend,
-                config_providers=MagicMock(),
+                config_providers=_DEFAULT_PROVIDERS,
             )
         assert result is not None
         parsed = json.loads(result)
@@ -153,51 +155,51 @@ class TestCheckDispatchFeasibilityUnit:
 
     def test_empty_post_prune_names_returns_none(self) -> None:
         """An empty post-prune step list (all pruned) returns None (1d)."""
-        from autoskillit.server.tools.tools_execution import _check_dispatch_feasibility
+        from autoskillit.server.tools._preflight import _check_dispatch_feasibility
 
         backend = _make_codex_backend()
         synthetic = _make_fix_required_hook()
-        with patch("autoskillit.hook_registry.HOOK_REGISTRY", [synthetic]):
+        with patch("autoskillit.server.tools._preflight.HOOK_REGISTRY", [synthetic]):
             result = _check_dispatch_feasibility(
                 post_prune_step_names=[],
                 active_recipe_steps={},
                 backend=backend,
-                config_providers=MagicMock(),
+                config_providers=_DEFAULT_PROVIDERS,
             )
         assert result is None
 
     def test_no_fix_required_hooks_returns_none(self) -> None:
         """A registry with no fix-required entries passes regardless of backend (1e)."""
-        from autoskillit.server.tools.tools_execution import _check_dispatch_feasibility
+        from autoskillit.server.tools._preflight import _check_dispatch_feasibility
 
         backend = _make_codex_backend()
         active_steps: dict[str, Any] = {
             "step1": _make_recipe_step("step1", tool="run_skill"),
         }
-        with patch("autoskillit.hook_registry.HOOK_REGISTRY", []):
+        with patch("autoskillit.server.tools._preflight.HOOK_REGISTRY", []):
             result = _check_dispatch_feasibility(
                 post_prune_step_names=["step1"],
                 active_recipe_steps=active_steps,
                 backend=backend,
-                config_providers=MagicMock(),
+                config_providers=_DEFAULT_PROVIDERS,
             )
         assert result is None
 
     def test_dormancy_synthetic_hook_always_blocks(self) -> None:
         """Synthetic fix-required hook permanently exercises failure path (1f)."""
-        from autoskillit.server.tools.tools_execution import _check_dispatch_feasibility
+        from autoskillit.server.tools._preflight import _check_dispatch_feasibility
 
         backend = _make_codex_backend()
         active_steps: dict[str, Any] = {
             "step1": _make_recipe_step("step1", tool="run_skill"),
         }
         dormancy = _make_dormancy_hook()
-        with patch("autoskillit.hook_registry.HOOK_REGISTRY", [dormancy]):
+        with patch("autoskillit.server.tools._preflight.HOOK_REGISTRY", [dormancy]):
             result = _check_dispatch_feasibility(
                 post_prune_step_names=["step1"],
                 active_recipe_steps=active_steps,
                 backend=backend,
-                config_providers=MagicMock(),
+                config_providers=_DEFAULT_PROVIDERS,
             )
         assert result is not None
         parsed = json.loads(result)
@@ -208,29 +210,21 @@ class TestCheckDispatchFeasibilityUnit:
     def test_provider_override_excludes_step(self) -> None:
         """A run_skill step with a provider profile that sets ANTHROPIC_BASE_URL
         on a non-anthropic backend is excluded from the check (1h)."""
-        from autoskillit.config._config_dataclasses import (
-            ProviderProfileDef,
-            ProvidersConfig,
-        )
-        from autoskillit.server.tools.tools_execution import _check_dispatch_feasibility
+        from autoskillit.config._config_dataclasses import ProvidersConfig
+        from autoskillit.server.tools._preflight import _check_dispatch_feasibility
 
         backend = _make_codex_backend()
         active_steps: dict[str, Any] = {
             "step1": _make_recipe_step("step1", tool="run_skill", provider="myprofile"),
         }
-        profile = ProviderProfileDef(
-            name="myprofile",
-            base_url="https://example.com",
-        )
         providers_config = ProvidersConfig(
-            profiles={"myprofile": profile},
-            resolved_profiles={"myprofile": profile},
+            profiles={"myprofile": {"base_url": "https://example.com"}},
             recipe_overrides={},
             step_overrides={},
             default_provider="myprofile",
         )
         synthetic = _make_fix_required_hook()
-        with patch("autoskillit.hook_registry.HOOK_REGISTRY", [synthetic]):
+        with patch("autoskillit.server.tools._preflight.HOOK_REGISTRY", [synthetic]):
             result = _check_dispatch_feasibility(
                 post_prune_step_names=["step1"],
                 active_recipe_steps=active_steps,
@@ -241,37 +235,37 @@ class TestCheckDispatchFeasibilityUnit:
 
     def test_no_backend_returns_none(self) -> None:
         """When backend is None, the preflight returns None (no check possible)."""
-        from autoskillit.server.tools.tools_execution import _check_dispatch_feasibility
+        from autoskillit.server.tools._preflight import _check_dispatch_feasibility
 
         backend = None
         active_steps: dict[str, Any] = {
             "step1": _make_recipe_step("step1", tool="run_skill"),
         }
         synthetic = _make_fix_required_hook()
-        with patch("autoskillit.hook_registry.HOOK_REGISTRY", [synthetic]):
+        with patch("autoskillit.server.tools._preflight.HOOK_REGISTRY", [synthetic]):
             result = _check_dispatch_feasibility(
                 post_prune_step_names=["step1"],
                 active_recipe_steps=active_steps,
                 backend=backend,
-                config_providers=MagicMock(),
+                config_providers=_DEFAULT_PROVIDERS,
             )
         assert result is None
 
     def test_error_includes_escape_hatch(self) -> None:
         """Preflight error envelope mentions the provider-override escape hatch."""
-        from autoskillit.server.tools.tools_execution import _check_dispatch_feasibility
+        from autoskillit.server.tools._preflight import _check_dispatch_feasibility
 
         backend = _make_codex_backend()
         active_steps: dict[str, Any] = {
             "step1": _make_recipe_step("step1", tool="run_skill"),
         }
         synthetic = _make_fix_required_hook()
-        with patch("autoskillit.hook_registry.HOOK_REGISTRY", [synthetic]):
+        with patch("autoskillit.server.tools._preflight.HOOK_REGISTRY", [synthetic]):
             result = _check_dispatch_feasibility(
                 post_prune_step_names=["step1"],
                 active_recipe_steps=active_steps,
                 backend=backend,
-                config_providers=MagicMock(),
+                config_providers=_DEFAULT_PROVIDERS,
             )
         assert result is not None
         parsed = json.loads(result)
@@ -288,7 +282,7 @@ class TestPreflightWiring:
 
     def test_check_dispatch_feasibility_is_importable(self) -> None:
         """The shared function exists and is importable."""
-        from autoskillit.server.tools.tools_execution import _check_dispatch_feasibility
+        from autoskillit.server.tools._preflight import _check_dispatch_feasibility
 
         assert callable(_check_dispatch_feasibility)
